Move the DoesNothingAi to the ai package
The DoesNothingAi label is defined in PlayerTypes because it is used in
the player selection to pick a default AI for certain types of players.
