+++ ai/src/main/java/org/triplea/ai/doesnothing/DoesNothingAi.java
@@ -1,4 +1,4 @@
-package games.strategy.triplea.ai.weak;
+package org.triplea.ai.doesnothing;
 
 import games.strategy.engine.data.Change;
 import games.strategy.engine.data.GameData;
@@ -7,7 +7,8 @@ import games.strategy.engine.data.ResourceCollection;
 import games.strategy.engine.data.changefactory.ChangeFactory;
 import games.strategy.engine.delegate.IDelegateBridge;
 import games.strategy.engine.framework.startup.ui.PlayerTypes;
-import games.strategy.triplea.ai.AbstractBuiltInAi;
+import games.strategy.triplea.ai.AbstractAi;
+import games.strategy.triplea.ai.weak.WeakAi;
 import games.strategy.triplea.delegate.remote.IAbstractForumPosterDelegate;
 import games.strategy.triplea.delegate.remote.IAbstractPlaceDelegate;
 import games.strategy.triplea.delegate.remote.IMoveDelegate;
@@ -18,15 +19,10 @@ import games.strategy.triplea.delegate.remote.ITechDelegate;
  * An AI implementation that takes no action except to purchase and place units according to very
  * simple rules.
  */
-public class DoesNothingAi extends AbstractBuiltInAi {
+public class DoesNothingAi extends AbstractAi {
 
-  public DoesNothingAi(final String name) {
-    super(name);
-  }
-
-  @Override
-  public PlayerTypes.Type getPlayerType() {
-    return PlayerTypes.DOES_NOTHING_AI;
+  public DoesNothingAi(final String name, final PlayerTypes.AiType playerType) {
+    super(name, playerType);
   }
 
   @Override
