+++ ai/src/main/resources/META-INF/services/games.strategy.triplea.ai.AiProvider
@@ -1,1 +1,2 @@
+org.triplea.ai.doesnothing.DoesNothingAiProvider
 org.triplea.ai.flowfield.FlowFieldAiProvider
